Create dedicated  file for templates rendering specific functions #49
