+++ config/app.php
@@ -72,18 +72,19 @@ return [
     */
     'autoload' => [
         'helpers.php',
-        'Structure/navs.php',
-        'Structure/widgets.php',
-        'Structure/sidebars.php',
-        'Structure/posttypes.php',
-        'Structure/taxonomies.php',
-        'Structure/shortcodes.php',
-        'Structure/thumbnails.php',
+        'Http/assets.php',
+        'Http/ajaxes.php',
         'Setup/actions.php',
         'Setup/filters.php',
         'Setup/supports.php',
         'Setup/services.php',
-        'Http/assets.php',
-        'Http/ajaxes.php',
+        'Structure/navs.php',
+        'Structure/widgets.php',
+        'Structure/sidebars.php',
+        'Structure/templates.php',
+        'Structure/posttypes.php',
+        'Structure/taxonomies.php',
+        'Structure/shortcodes.php',
+        'Structure/thumbnails.php',
     ],
 ];
